doi thu vien echo va update them folder service, repository, controler

+++ models/meta.go
@@ -1,7 +1,7 @@
 package models
 
 type Meta struct {
-	Page  int `json:"page"`
-	Limit int `json:"limit"`
-	Total int `json:"total"`
+	Page  int   `json:"page"`
+	Limit int   `json:"limit"`
+	Total int64 `json:"total"`
 }
